Refactor: Modularize app.js for clarity and maintainability
This commit addresses the first item on the project roadmap:
"Iteration 1: Refactor for Clarity and Modularity".

The main `app.js` file has been broken down into several focused modules:

- `webgl-setup.js`: Handles initial WebGL context setup.
- `shader-utils.js`: Provides utilities for shader compilation and program linking.
- `spoon-model.js`: Contains the geometry data (vertices, normals, indices) for the spoon object.
- `grid-model.js`: Contains the logic for generating vertex data for the grid.
- `renderer.js`: Manages the rendering loop, drawing of objects, and animation state.

`app.js` now acts as the main entry point, orchestrating the initialization and execution by importing functionalities from these new modules.
`index.html` has been updated to load `app.js` as an ES6 module.

This refactoring improves code readability, makes the project easier to maintain, and facilitates the implementation of future features.
The README.md file has been updated to mark this iteration as complete.
