+++ index.html
@@ -140,6 +140,6 @@
 </head>
 <body>
   <canvas id="glCanvas"></canvas>
-  <script src="app.js"></script>
+  <script type="module" src="app.js"></script>
 </body>
 </html>
